Always delete any associated disabled module entries
This needs to be done due to the nature of checkboxes in html forms. If
the user unchecks every checkbox, then the form sends no information
regarding the checkboxes which means that our check against
'disabled_modules' in the form data will fail and the previously
selected modules will still be "selected/saved".